+++ .github/workflows/maven.yml
@@ -26,7 +26,7 @@ jobs:
       env:
         API_KEY: ${{ secrets.API_KEY }}
       run: |
-        touch src/main/resources/fi/tuni/ec/backend/controller/config.properties && printf "entsoe_api_key=$API_KEY" >> src/main/resources/fi/tuni/ec/backend/controller/config.properties
+        touch src/main/resources/fi/tuni/ec/backend/api/config.properties && printf "entsoe_api_key=$API_KEY" >> src/main/resources/fi/tuni/ec/backend/api/config.properties
       
     - name: Set up JDK 22
       uses: actions/setup-java@v4
