+++ src/main/java/fi/tuni/ec/backend/controller/MainController.java
@@ -4,6 +4,8 @@ import java.io.IOException;
 // import java.time.LocalDate;
 // import java.time.YearMonth;
 import java.util.List;
+
+import fi.tuni.ec.api.ApiService;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.image.Image;
